+++ test/e2e/steps/has_group.dart
@@ -65,6 +65,31 @@ final StepDefinitionGeneric hasGroupNamed =
         ..timeout = const Duration(minutes: 5),
     );
 
+/// Creates a [Chat]-group with the the provided [TestUser].
+///
+/// Examples:
+/// - Given Alice has "Name" group.
+final StepDefinitionGeneric hasGroupNamedInArchive =
+    given2<TestUser, String, CustomWorld>(
+      '{user} has {string} group in archive\$',
+      (TestUser user, String name, context) async {
+        final provider = GraphQlProvider();
+        provider.token = context.world.sessions[user.name]?.token;
+
+        final ChatMixin chat = await provider.createGroupChat(
+          [],
+          name: ChatName(name),
+        );
+
+        await provider.toggleChatArchivation(chat.id, true);
+
+        context.world.groups[name] = chat.id;
+        provider.disconnect();
+      },
+      configuration: StepDefinitionConfiguration()
+        ..timeout = const Duration(minutes: 5),
+    );
+
 /// Creates a [Chat]-group with the provided [User] and the authenticated
 /// [MyUser].
 ///