+++ test/unit/call_test.dart
@@ -654,6 +654,7 @@ class _FakeGraphQlProvider extends MockedGraphQlProvider {
   Stream<QueryResult> recentChatsTopEvents(
     int count, {
     bool noFavorite = false,
+    bool archived = false,
     bool? withOngoingCalls,
   }) => Stream.value(
     QueryResult.internal(
@@ -772,6 +773,7 @@ class _FakeGraphQlProvider extends MockedGraphQlProvider {
     int? last,
     RecentChatsCursor? before,
     bool noFavorite = false,
+    bool archived = false,
     bool? withOngoingCalls,
   }) async {
     return RecentChats$Query.fromJson(chatsQuery);
@@ -940,6 +942,7 @@ final chatData = {
   'members': {'nodes': [], 'totalCount': 0},
   'kind': 'GROUP',
   'isHidden': false,
+  'isArchived': false,
   'muted': null,
   'directLink': null,
   'createdAt': '2021-12-15T15:11:18.316846+00:00',